Воссоединение с ответвлением, после разработки новой Админ-панели ------------- [new]   Новая Админ панель [dev]   По новому организован уровень доступа, удалены старые переменные, осталась одна $rights [dev]   Связанные доработки во всех подсистемах

+++ str/guest.php
@@ -22,41 +22,37 @@ require_once ("../incfiles/core.php");
 require_once ("../incfiles/head.php");
 
 // Проверяем права доступа в Админ-Клуб
-if (isset($_SESSION['ga']) && $dostmod != 1)
-    unset($_SESSION['ga']);
+if (isset ($_SESSION['ga']) && $rights < 1)
+    unset ($_SESSION['ga']);
 
 // Задаем заголовки страницы
-$textl = isset($_SESSION['ga']) ? 'Админ-Клуб' : 'Гостевая';
+$textl = isset ($_SESSION['ga']) ? 'Админ-Клуб' : 'Гостевая';
 
 // Если гостевая закрыта, выводим сообщение и закрываем доступ (кроме Админов)
-if (!$set['mod_guest'] && !$dostadm)
-{
+if (!$set['mod_guest'] && $rights < 7) {
     echo '<div class="rmenu"><p>Гостевая закрыта</p></div>';
     require_once ("../incfiles/end.php");
     exit;
 }
 
-switch ($act)
-{
-    case "delpost":
+switch ($act) {
+    case "delpost" :
         ////////////////////////////////////////////////////////////
         // Удаление отдельного поста                              //
         ////////////////////////////////////////////////////////////
-        if ($dostsmod && $id)
-        {
-            if (isset($_GET['yes']))
-            {
+        if ($rights >= 6 && $id) {
+            if (isset ($_GET['yes'])) {
                 mysql_query("DELETE FROM `guest` WHERE `id`='" . $id . "' LIMIT 1");
                 header("Location: guest.php");
-            } else
-            {
+            }
+            else {
                 echo '<p>Вы действительно хотите удалить пост?<br/>';
                 echo '<a href="guest.php?act=delpost&amp;id=' . $id . '&amp;yes">Удалить</a> | <a href="guest.php">Отмена</a></p>';
             }
         }
         break;
 
-    case "trans":
+    case "trans" :
         ////////////////////////////////////////////////////////////
         // Справка по транслиту                                   //
         ////////////////////////////////////////////////////////////
@@ -64,75 +60,65 @@ switch ($act)
         echo '<br/><br/><a href="' . htmlspecialchars(getenv("HTTP_REFERER")) . '">Назад</a><br/>';
         break;
 
-    case "say":
+    case "say" :
         ////////////////////////////////////////////////////////////
         // Добавление нового поста                                //
         ////////////////////////////////////////////////////////////
-        if (empty($user_id) && empty($_POST['name']))
-        {
+        if (empty ($user_id) && empty ($_POST['name'])) {
             echo "<p>Вы не ввели имя!<br/><a href='guest.php'>Назад</a></p>";
             require_once ("../incfiles/end.php");
             exit;
         }
-        if (empty($_POST['msg']))
-        {
+        if (empty ($_POST['msg'])) {
             echo "<p>Вы не ввели сообщение!<br/><a href='guest.php'>Назад</a></p>";
             require_once ("../incfiles/end.php");
             exit;
         }
-        if (empty($_SESSION['guest']) || $ban['1'] || $ban['13'])
-        {
+        if (empty ($_SESSION['guest']) || $ban['1'] || $ban['13']) {
             echo "<p><b>Спам!</b></p>";
             require_once ("../incfiles/end.php");
             exit;
         }
-        if (!$user_id && $_SESSION['code'] != $_POST['code'])
-        {
+        if (!$user_id && $_SESSION['code'] != $_POST['code']) {
             echo "<p>Код введен неверно!<br/><a href='guest.php'>Назад</a></p>";
             require_once ("../incfiles/end.php");
             exit;
         }
         // Задаем куда вставляем, в Админ клуб (1), или в Гастивуху (0)
-        $admset = isset($_SESSION['ga']) ? 1:
-        0;
+        $admset = isset ($_SESSION['ga']) ? 1 : 0;
         // Антиспам, проверка на частоту добавления сообщений
-        if ($user_id)
-        {
-            $old = ($rights > 0 || $dostsadm = 1) ? 10 : 30;
+        if ($user_id) {
+            $old = ($rights > 0) ? 10 : 30;
             $spam = $datauser['lastpost'] > ($realtime - $old) ? 1 : false;
-        } else
-        {
+        }
+        else {
             $req = mysql_query("SELECT COUNT(*) FROM `guest` WHERE `soft`='" . mysql_real_escape_string($agn) . "' AND `time` >='" . ($realtime - 30) . "' AND `ip` ='" . $ipl . "' AND `adm`='" . $admset . "'");
             $spam = mysql_result($req, 0) > 0 ? 1 : false;
         }
-        if ($spam)
-        {
+        if ($spam) {
             echo "<p><b>Антифлуд!</b><br />Вы не можете так часто добавлять сообщения<br/>Порог $old секунд<br/><br/><a href='guest.php'>Назад</a></p>";
             require_once ("../incfiles/end.php");
             exit;
         }
-        unset($_SESSION['guest']);
+        unset ($_SESSION['guest']);
         $_SESSION['code'] = rand(1000, 9999);
         $name = mb_substr(trim($_POST['name']), 0, 20);
-        if ($user_id)
-        {
+        if ($user_id) {
             $from = $login;
-        } else
-        {
+        }
+        else {
             $from = check($name);
             $user_id = 0;
         }
         $msg = trim($_POST['msg']);
         $msg = mb_substr($msg, 0, 500);
-        if ($_POST['msgtrans'] == 1)
-        {
+        if ($_POST['msgtrans'] == 1) {
             $msg = trans($msg);
         }
         // Проверка на одинаковые сообщения
         $req = mysql_query("SELECT * FROM `guest` WHERE `user_id` = '$user_id' ORDER BY `time` DESC");
         $res = mysql_fetch_array($req);
-        if ($res['text'] == $msg)
-        {
+        if ($res['text'] == $msg) {
             header("location: guest.php");
             exit;
         }
@@ -146,22 +132,19 @@ switch ($act)
 		`ip`='" . $ipl . "',
 		`browser`='" . mysql_real_escape_string($agn) . "'");
         // Фиксируем время последнего поста (антиспам)
-        if ($user_id)
-        {
+        if ($user_id) {
             $postguest = $datauser['postguest'] + 1;
             mysql_query("UPDATE `users` SET `postguest` = '$postguest', `lastpost` = '" . $realtime . "' WHERE `id` = '" . $user_id . "'");
         }
         header("location: guest.php");
         break;
 
-    case "otvet":
+    case "otvet" :
         ////////////////////////////////////////////////////////////
         // Добавление "ответа Админа"                             //
         ////////////////////////////////////////////////////////////
-        if ($dostsmod && $id)
-        {
-            if (isset($_POST['submit']))
-            {
+        if ($rights >= 6 && $id) {
+            if (isset ($_POST['submit'])) {
                 $otv = mb_substr($_POST['otv'], 0, 500);
                 mysql_query("UPDATE `guest` SET
 				`admin` = '" . $login . "',
@@ -169,12 +152,11 @@ switch ($act)
 				`otime` = '" . $realtime . "'
 				WHERE `id` = '" . $id . "'");
                 header("location: guest.php");
-            } else
-            {
+            }
+            else {
                 $ps = mysql_query("select * from `guest` where id='" . $id . "'");
                 $ps1 = mysql_fetch_array($ps);
-                if (!empty($ps1['otvet']))
-                {
+                if (!empty ($ps1['otvet'])) {
                     echo "<br /><b>Внимание!<br />На этот пост уже ответили.</b><br/><br/>";
                 }
                 $text = htmlentities($ps1['text'], ENT_QUOTES, 'UTF-8');
@@ -184,14 +166,12 @@ switch ($act)
         }
         break;
 
-    case "edit":
+    case "edit" :
         ////////////////////////////////////////////////////////////
         // Редактирование поста                                   //
         ////////////////////////////////////////////////////////////
-        if ($dostsmod && $id)
-        {
-            if (isset($_POST['submit']))
-            {
+        if ($rights >= 6 && $id) {
+            if (isset ($_POST['submit'])) {
                 $req = mysql_query("SELECT `edit_count` FROM `guest` WHERE `id`='" . $id . "' LIMIT 1");
                 $res = mysql_fetch_array($req);
                 $edit_count = $res['edit_count'] + 1;
@@ -203,8 +183,8 @@ switch ($act)
 				`edit_count`='" . $edit_count . "'
 				WHERE `id`='" . $id . "'");
                 header("location: guest.php");
-            } else
-            {
+            }
+            else {
                 $req = mysql_query("SELECT * FROM `guest` WHERE `id` = '" . $id . "' LIMIT 1");
                 $res = mysql_fetch_array($req);
                 $text = htmlentities($res['text'], ENT_QUOTES, 'UTF-8');
@@ -218,38 +198,35 @@ switch ($act)
         }
         break;
 
-    case 'clean':
+    case 'clean' :
         ////////////////////////////////////////////////////////////
         // Очистка Гостевой                                       //
         ////////////////////////////////////////////////////////////
-        if ($dostadm)
-        {
-            if (isset($_POST['submit']))
-            {
-                $adm = isset($_SESSION['ga']) ? 1 : 0;
-                $cl = isset($_POST['cl']) ? intval($_POST['cl']) : '';
-                switch ($cl)
-                {
-                    case '1':
+        if ($rights >= 7) {
+            if (isset ($_POST['submit'])) {
+                $adm = isset ($_SESSION['ga']) ? 1 : 0;
+                $cl = isset ($_POST['cl']) ? intval($_POST['cl']) : '';
+                switch ($cl) {
+                    case '1' :
                         // Чистим сообщения, старше 1 дня
                         mysql_query("DELETE FROM `guest` WHERE `adm`='$adm' AND `time` < '" . ($realtime - 86400) . "'");
                         echo '<p>Удалены все сообщения, старше 1 дня.</p><p><a href="guest.php">Вернуться</a></p>';
                         break;
 
-                    case '2':
+                    case '2' :
                         // Проводим полную очистку
                         mysql_query("DELETE FROM `guest` WHERE `adm`='$adm'");
                         echo '<p>Удалены все сообщения.</p><p><a href="guest.php">Вернуться</a></p>';
                         break;
 
-                    default:
+                    default :
                         // Чистим сообщения, старше 1 недели
                         mysql_query("DELETE FROM `guest` WHERE `adm`='$adm' AND `time`<='" . ($realtime - 604800) . "';");
                         echo '<p>Все сообщения, старше 1 недели удалены из Гостевой.</p><p><a href="guest.php">В Гостевую</a></p>';
                 }
                 mysql_query("OPTIMIZE TABLE `guest`");
-            } else
-            {
+            }
+            else {
                 echo '<p><b>Очистка сообщений</b></p>';
                 echo '<u>Что чистим?</u>';
                 echo '<form id="clean" method="post" action="guest.php?act=clean">';
@@ -263,24 +240,22 @@ switch ($act)
         }
         break;
 
-    case 'ga':
+    case 'ga' :
         ////////////////////////////////////////////////////////////
         // Переключение режима работы Гостевая / Админ-клуб       //
         ////////////////////////////////////////////////////////////
-        if ($dostmod == 1)
-        {
-            if ($_GET['do'] == 'set')
-            {
+        if ($rights >= 1) {
+            if ($_GET['do'] == 'set') {
                 $_SESSION['ga'] = 1;
                 $textl = 'Админ-Клуб';
-            } else
-            {
-                unset($_SESSION['ga']);
+            }
+            else {
+                unset ($_SESSION['ga']);
                 $textl = 'Гостевая';
             }
         }
 
-    default:
+    default :
         ////////////////////////////////////////////////////////////
         // Отображаем Гостевую, или Админ клуб                    //
         ////////////////////////////////////////////////////////////
@@ -288,57 +263,49 @@ switch ($act)
             echo '<p><span class="red"><b>Гостевая закрыта!</b></span></p>';
         echo '<div class="phdr"><b>Гостевая</b></div>';
         // Форма ввода нового сообщения
-        if (($user_id || $set['mod_guest'] == 2) && !$ban['1'] && !$ban['13'])
-        {
+        if (($user_id || $set['mod_guest'] == 2) && !$ban['1'] && !$ban['13']) {
             $_SESSION['guest'] = rand(1000, 9999);
             echo '<div class="gmenu"><form action="guest.php?act=say" method="post">';
-            if (!$user_id)
-            {
+            if (!$user_id) {
                 echo "Имя(max. 25):<br/><input type='text' name='name' maxlength='25'/><br/>";
             }
             echo 'Сообщение(max. 500):<br/><textarea cols="20" rows="2" name="msg"></textarea><br/>';
             if ($set_user['translit'])
                 echo '<input type="checkbox" name="msgtrans" value="1" /> Транслит сообщения<br/>';
-            if (!$user_id)
-            {
+            if (!$user_id) {
                 // CAPTCHA для гостей
                 $_SESSION['code'] = rand(1000, 9999);
                 echo '<img src="../code.php" alt="Код"/><br />';
                 echo '<input type="text" size="4" maxlength="4"  name="code"/>&nbsp;введите код<br />';
             }
             echo "<input type='submit' title='Нажмите для отправки' name='submit' value='Отправить'/></form></div>";
-        } else
-        {
+        }
+        else {
             echo '<div class="rmenu">Писать могут только <a href="../in.php">авторизованные</a> посетители</div>';
         }
-        if (isset($_SESSION['ga']) && ($login == $nickadmina || $login == $nickadmina2 || $rights >= "1"))
-        {
+        if (isset ($_SESSION['ga']) && ($login == $nickadmina || $login == $nickadmina2 || $rights >= "1")) {
             $req = mysql_query("SELECT COUNT(*) FROM `guest` WHERE `adm`='1'");
-        } else
-        {
+        }
+        else {
             $req = mysql_query("SELECT COUNT(*) FROM `guest` WHERE `adm`='0'");
         }
-        $colmes = mysql_result($req, 0); // Число сообщений в гастивухе
-        if ($colmes > 0)
-        {
-            if (isset($_SESSION['ga']) && ($login == $nickadmina || $login == $nickadmina2 || $rights >= "1"))
-            {
+        $colmes = mysql_result($req, 0);        // Число сообщений в гастивухе
+        if ($colmes > 0) {
+            if (isset ($_SESSION['ga']) && ($login == $nickadmina || $login == $nickadmina2 || $rights >= "1")) {
                 // Запрос для Админ клуба
                 echo '<div class="rmenu"><b>АДМИН-КЛУБ</b></div>';
                 $req = mysql_query("SELECT `guest`.*, `guest`.`id` AS `gid`, `users`.`rights`, `users`.`lastdate`, `users`.`sex`, `users`.`status`, `users`.`datereg`, `users`.`id`
 				FROM `guest` LEFT JOIN `users` ON `guest`.`user_id` = `users`.`id` WHERE `guest`.`adm`='1' ORDER BY `time` DESC LIMIT " . $start . "," . $kmess);
-            } else
-            {
+            }
+            else {
                 // Запрос для обычной Гастивухи
                 $req = mysql_query("SELECT `guest`.*, `guest`.`id` AS `gid`, `users`.`rights`, `users`.`lastdate`, `users`.`sex`, `users`.`status`, `users`.`datereg`, `users`.`id`
 				FROM `guest` LEFT JOIN `users` ON `guest`.`user_id` = `users`.`id` WHERE `guest`.`adm`='0' ORDER BY `time` DESC LIMIT " . $start . "," . $kmess);
             }
-            while ($res = mysql_fetch_assoc($req))
-            {
+            while ($res = mysql_fetch_assoc($req)) {
                 $text = '';
                 echo ($i % 2) ? '<div class="list2">' : '<div class="list1">';
-                if (empty($res['id']))
-                {
+                if (empty ($res['id'])) {
                     // Запрос по гостям
                     $req_g = mysql_query("SELECT `lastdate` FROM `cms_guests` WHERE `session_id` = '" . md5($res['ip'] . $res['browser']) . "' LIMIT 1");
                     $res_g = mysql_fetch_assoc($req_g);
@@ -346,25 +313,22 @@ switch ($act)
                 }
                 // Время создания поста
                 $text = ' <span class="gray">(' . date("d.m.y / H:i", $res['time'] + $set_user['sdvig'] * 3600) . ')</span>';
-                if ($res['user_id'])
-                {
+                if ($res['user_id']) {
                     // Для зарегистрированных показываем ссылки и смайлы
                     $post = checkout($res['text'], 1, 1);
                     if ($set_user['smileys'])
                         $post = smileys($post, $res['rights'] >= 1 ? 1 : 0);
-                } else
-                {
+                }
+                else {
                     // Для гостей фильтруем ссылки
                     $post = antilink(checkout($res['text'], 0, 2));
                 }
-                if ($res['edit_count'])
-                {
+                if ($res['edit_count']) {
                     // Если пост редактировался, показываем кем и когда
                     $dizm = date("d.m /H:i", $res['edit_time'] + $set_user['sdvig'] * 3600);
                     $post .= '<br /><span class="gray"><small>Изм. <b>' . $res['edit_who'] . '</b> (' . $dizm . ') <b>[' . $res['edit_count'] . ']</b></small></span>';
                 }
-                if (!empty($res['otvet']))
-                {
+                if (!empty ($res['otvet'])) {
                     // Ответ Администрации
                     $otvet = checkout($res['otvet'], 1, 1);
                     $vrp1 = $res['otime'] + $set_user['sdvig'] * 3600;
@@ -373,29 +337,28 @@ switch ($act)
                         $otvet = smileys($otvet, 1);
                     $post .= '<div class="reply"><b>' . $res['admin'] . '</b>: (' . $vr1 . ')<br/>' . $otvet . '</div>';
                 }
-                $subtext = '<a href="guest.php?act=otvet&amp;id=' . $res['gid'] . '">Ответить</a> | <a href="guest.php?act=edit&amp;id=' . $res['gid'] . '">Изменить</a> | <a href="guest.php?act=delpost&amp;id=' . $res['gid'] . '">Удалить</a>';
-                echo show_user($res, 1, ($dostsmod ? 1 : 0), $text, $post, ($dostsmod ? $subtext : ''));
+                $subtext = '<a href="guest.php?act=otvet&amp;id=' . $res['gid'] . '">Ответить</a>' . ($rights >= 6 && $rights >= $res['rights'] ? ' | <a href="guest.php?act=edit&amp;id=' . $res['gid'] . '">Изменить</a> | <a href="guest.php?act=delpost&amp;id=' . $res['gid'] . '">Удалить</a>' : '');
+                echo show_user($res, 1, ($rights >= 6 && $rights >= $res['rights'] ? 1 : 0), $text, $post, ($rights >= 6 ? $subtext : ''));
                 echo '</div>';
                 ++$i;
             }
             echo '<div class="phdr">Всего сообщений: ' . $colmes . '</div>';
-            if ($colmes > $kmess)
-            {
+            if ($colmes > $kmess) {
                 echo '<p>' . pagenav('guest.php?', $start, $colmes, $kmess) . '</p>';
                 echo '<p><form action="guest.php" method="get"><input type="text" name="page" size="2"/><input type="submit" value="К странице &gt;&gt;"/></form></p>';
             }
             echo '<p><div class="func">';
             // Для Админов даем ссылку на чистку Гостевой
-            if ($dostadm)
+            if ($rights >= 7)
                 echo '<a href="guest.php?act=clean">Чистка истории</a><br />';
             echo '</div></p>';
-        } else
-        {
+        }
+        else {
             echo '<p>В Гостевой сообщений нет.</p>';
         }
         // Ссылка на Админ-клуб
-        if ($dostmod)
-            echo (isset($_SESSION['ga']) ? '<p><a href="guest.php?act=ga"><b>Гостевая &gt;&gt;</b></a></p>' : '<p><a href="guest.php?act=ga&amp;do=set"><b>Админ-Клуб &gt;&gt;</b></a></p>');
+        if ($rights >= 1)
+            echo (isset ($_SESSION['ga']) ? '<p><a href="guest.php?act=ga"><b>Гостевая &gt;&gt;</b></a></p>' : '<p><a href="guest.php?act=ga&amp;do=set"><b>Админ-Клуб &gt;&gt;</b></a></p>');
         break;
 }
